+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for CLEXI - Client Extension Interface
 
+## v0.9.1
+* Added package 'usb' to interface with USB devices
+* Added custom gpio-item example 'respeaker-usb-array-v2' to control the LED array of a ReSpeaker Mic Array v2.0 via USB
+* Added 'test(options)' to GPIO interface, e.g. for local testing
+
 ## v0.9.0
 * Switched to Node.js 14 as officially supported version and changed repository for 'node-beacon-scanner' to more up-to-date version
 * Added new extension 'gpio-interface' to communicate with the GPIO pins of Raspberry Pi (and similar boards)